+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/state/ContentState.kt
@@ -33,6 +33,8 @@ import mozilla.components.concept.engine.window.WindowRequest
  * @property findResults the list of results of the latest "find in page" operation.
  * @property windowRequest the last received [WindowRequest].
  * @property searchRequest the last received [SearchRequest]
+ * @property fullScreen true if the page is full screen, false if not.
+ * @property layoutInDisplayCutoutMode the display layout cutout mode state.
  */
 data class ContentState(
     val url: String,
@@ -49,5 +51,7 @@ data class ContentState(
     val promptRequest: PromptRequest? = null,
     val findResults: List<FindResultState> = emptyList(),
     val windowRequest: WindowRequest? = null,
-    val searchRequest: SearchRequest? = null
+    val searchRequest: SearchRequest? = null,
+    val fullScreen: Boolean = false,
+    val layoutInDisplayCutoutMode: Int = 0
 )