+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/FullScreenFeatureTest.kt
@@ -94,7 +94,7 @@ class FullScreenFeatureTest {
         sessionUseCases: SessionUseCases,
         sessionId: String? = null,
         fullScreenChanged: (Boolean) -> Unit
-    ) : FullScreenFeature(sessionManager, sessionUseCases, sessionId, fullScreenChanged) {
+    ) : FullScreenFeature(sessionManager, sessionUseCases, sessionId, fullScreenChanged = fullScreenChanged) {
         public override var activeSession: Session? = null
     }
 }